python/svm/requirements.txt
aiohappyeyeballs==2.4.8
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
certifi==2025.1.31
deprecation==2.1.0
frozenlist==1.5.0
gotrue==2.11.4
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
imbalanced-learn==0.13.0
joblib==1.4.2
multidict==6.1.0
numpy==2.2.4
packaging==24.2
pandas==2.2.3
postgrest==0.19.3
propcache==0.3.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
realtime==2.4.2
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sklearn-compat==0.1.3
sniffio==1.3.1
storage3==0.11.3
StrEnum==0.4.15
supabase==2.13.0
supafunc==0.9.4
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2025.2
websockets==14.2
yarl==1.18.3
